Refactored Static methods into a single static class
