+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/udp/AbstractUdpReceiverPublisher.java
@@ -1,17 +1,13 @@
 package us.dot.its.jpo.ode.udp;
 
 import java.net.DatagramSocket;
-import java.net.DatagramPacket;
 import java.net.SocketException;
 
-import org.apache.tomcat.util.buf.HexUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import us.dot.its.jpo.ode.OdeProperties;
-import us.dot.its.jpo.ode.model.OdeAsn1Payload;
-import us.dot.its.jpo.ode.uper.UperUtil;
 
 public abstract class AbstractUdpReceiverPublisher implements Runnable {
 
@@ -57,24 +53,6 @@ public abstract class AbstractUdpReceiverPublisher implements Runnable {
       }
    }
 
-   public static OdeAsn1Payload getPayloadHexString(DatagramPacket packet, UperUtil.SupportedMessageTypes msgType) {
-      String startFlag = UperUtil.getStartFlag(msgType);
-      // extract the actual packet from the buffer
-      byte[] payload = packet.getData();
-      if (payload == null)
-         return null;
-      // convert bytes to hex string and verify identity
-      String payloadHexString = HexUtils.toHexString(payload).toLowerCase();
-      if (payloadHexString.indexOf(startFlag) == -1)
-         return null;
-      
-      logger.debug("Full {} packet: {}", msgType, payloadHexString);
-      payloadHexString = UperUtil.stripDot3Header(payloadHexString, startFlag);
-      logger.debug("Stripped {} packet: {}", msgType, payloadHexString);
-
-      OdeAsn1Payload timPayload = new OdeAsn1Payload(HexUtils.fromHexString(payloadHexString));
-      
-      return timPayload;
-   }
+   
 
 }
