+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/udp/psm/PsmReceiver.java
@@ -1,26 +1,14 @@
 package us.dot.its.jpo.ode.udp.psm;
 
 import java.net.DatagramPacket;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import us.dot.its.jpo.ode.coder.StringPublisher;
-import us.dot.its.jpo.ode.model.OdeAsn1Data;
-import us.dot.its.jpo.ode.model.OdeAsn1Payload;
-import us.dot.its.jpo.ode.model.OdeLogMetadata.RecordType;
-import us.dot.its.jpo.ode.model.OdeLogMetadata.SecurityResultCode;
-import us.dot.its.jpo.ode.model.OdeMsgMetadata.GeneratedBy;
-import us.dot.its.jpo.ode.model.OdePsmMetadata.PsmSource;
-import us.dot.its.jpo.ode.model.OdePsmMetadata;
 import us.dot.its.jpo.ode.OdeProperties;
 import us.dot.its.jpo.ode.udp.AbstractUdpReceiverPublisher;
-import us.dot.its.jpo.ode.uper.UperUtil;
-import us.dot.its.jpo.ode.util.JsonUtils;
+import us.dot.its.jpo.ode.udp.UdpHexDecoder;
 
 public class PsmReceiver extends AbstractUdpReceiverPublisher {
     private static Logger logger = LoggerFactory.getLogger(PsmReceiver.class);
@@ -54,7 +42,7 @@ public class PsmReceiver extends AbstractUdpReceiverPublisher {
                 logger.debug("Waiting for UDP PSM packets...");
                 socket.receive(packet);
                 if (packet.getLength() > 0) {
-                    String psmJson = buildJsonPsmFromPacket(packet);
+                    String psmJson = UdpHexDecoder.buildJsonPsmFromPacket(packet);
                     if(psmJson != null){
                         // Submit JSON to the OdeRawEncodedMessageJson Kafka Topic
                         psmPublisher.publish(psmJson, psmPublisher.getOdeProperties().getKafkaTopicOdeRawEncodedPSMJson());
@@ -66,29 +54,5 @@ public class PsmReceiver extends AbstractUdpReceiverPublisher {
         } while (!isStopped());
     }
 
-    public static String buildJsonPsmFromPacket(DatagramPacket packet){
-        String senderIp = packet.getAddress().getHostAddress();
-        int senderPort = packet.getPort();
-        logger.debug("Packet received from {}:{}", senderIp, senderPort);
-
-        // Create OdeMsgPayload and OdeLogMetadata objects and populate them
-        OdeAsn1Payload psmPayload = AbstractUdpReceiverPublisher.getPayloadHexString(packet, UperUtil.SupportedMessageTypes.PSM);
-        if (psmPayload == null)
-            return null;
-        OdePsmMetadata psmMetadata = new OdePsmMetadata(psmPayload);
-        // Add header data for the decoding process
-        ZonedDateTime utc = ZonedDateTime.now(ZoneOffset.UTC);
-        String timestamp = utc.format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"));
-        psmMetadata.setOdeReceivedAt(timestamp);
-
-        psmMetadata.setOriginIp(senderIp);
-        psmMetadata.setPsmSource(PsmSource.RSU);
-        psmMetadata.setRecordType(RecordType.psmTx);
-        psmMetadata.setRecordGeneratedBy(GeneratedBy.UNKNOWN);
-        psmMetadata.setSecurityResultCode(SecurityResultCode.success);
-
-        return JsonUtils.toJson(new OdeAsn1Data(psmMetadata, psmPayload), false);
-
-
-    }
+    
 }
