+++ jpo-ode-svcs/src/main/java/us/dot/its/jpo/ode/udp/srm/SrmReceiver.java
@@ -1,26 +1,14 @@
 package us.dot.its.jpo.ode.udp.srm;
 
 import java.net.DatagramPacket;
-import java.time.ZoneOffset;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
 import us.dot.its.jpo.ode.coder.StringPublisher;
-import us.dot.its.jpo.ode.model.OdeAsn1Data;
-import us.dot.its.jpo.ode.model.OdeAsn1Payload;
-import us.dot.its.jpo.ode.model.OdeLogMetadata.RecordType;
-import us.dot.its.jpo.ode.model.OdeLogMetadata.SecurityResultCode;
-import us.dot.its.jpo.ode.model.OdeMsgMetadata.GeneratedBy;
-import us.dot.its.jpo.ode.model.OdeSrmMetadata;
-import us.dot.its.jpo.ode.model.OdeSrmMetadata.SrmSource;
 import us.dot.its.jpo.ode.OdeProperties;
 import us.dot.its.jpo.ode.udp.AbstractUdpReceiverPublisher;
-import us.dot.its.jpo.ode.uper.UperUtil;
-import us.dot.its.jpo.ode.util.JsonUtils;
+import us.dot.its.jpo.ode.udp.UdpHexDecoder;
 
 public class SrmReceiver extends AbstractUdpReceiverPublisher {
     private static Logger logger = LoggerFactory.getLogger(SrmReceiver.class);
@@ -55,7 +43,7 @@ public class SrmReceiver extends AbstractUdpReceiverPublisher {
                 socket.receive(packet);
                 if (packet.getLength() > 0) {
                     
-                    String srmJson = buildJsonSrmFromPacket(packet);
+                    String srmJson = UdpHexDecoder.buildJsonSrmFromPacket(packet);
                     // Submit JSON to the OdeRawEncodedMessageJson Kafka Topic
                     srmPublisher.publish(srmJson, srmPublisher.getOdeProperties().getKafkaTopicOdeRawEncodedSRMJson());
                 }
@@ -65,29 +53,5 @@ public class SrmReceiver extends AbstractUdpReceiverPublisher {
         } while (!isStopped());
     }
 
-    public static String buildJsonSrmFromPacket(DatagramPacket packet){
-        String senderIp = packet.getAddress().getHostAddress();
-        int senderPort = packet.getPort();
-        logger.debug("Packet received from {}:{}", senderIp, senderPort);
-
-        // Create OdeMsgPayload and OdeLogMetadata objects and populate them
-        OdeAsn1Payload srmPayload = AbstractUdpReceiverPublisher.getPayloadHexString(packet, UperUtil.SupportedMessageTypes.SRM);
-        if (srmPayload == null)
-            return null;
-        OdeSrmMetadata srmMetadata = new OdeSrmMetadata(srmPayload);
-
-        // Add header data for the decoding process
-        ZonedDateTime utc = ZonedDateTime.now(ZoneOffset.UTC);
-        String timestamp = utc.format(DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS'Z'"));
-        srmMetadata.setOdeReceivedAt(timestamp);
-
-        srmMetadata.setOriginIp(senderIp);
-        srmMetadata.setSrmSource(SrmSource.RSU);
-        srmMetadata.setRecordType(RecordType.srmTx);
-        srmMetadata.setRecordGeneratedBy(GeneratedBy.OBU);
-        srmMetadata.setSecurityResultCode(SecurityResultCode.success);
-
-        return JsonUtils.toJson(new OdeAsn1Data(srmMetadata, srmPayload), false);
-
-    }
+    
 }
